Added parent class and changed pinch and grab detection to use this parent. Also moved the detectors to their own folder
