+++ Packages/Tracking/Core/Runtime/Scripts/ActionDetectors.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1e1244271cb14b241a402ca5159769bc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
