+++ Packages/Tracking/Core/Runtime/Scripts/ActionDetectors/ActionDetector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c7a3de12ef8a6074b86ada72e7c92b32
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
