+++ Packages/Tracking/Core/Runtime/Scripts/ActionDetectors/PinchDetector.cs.meta
